# library-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the back-end API for an online library management system",
        "title": "Library Management API 文档",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/borrow": {
            "post": {
                "description": "用户借阅图书，库存 -1，并创建借阅记录 (User borrows a book, reduces available copies, and creates borrowing record)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "借阅管理 (Borrow Management)"
                ],
                "summary": "借阅图书 (User borrows a book)",
                "parameters": [
                    {
                        "description": "借阅请求参数 (Borrow Request)",
                        "name": "borrowRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BorrowRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "借阅成功 (Success Response)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "请求数据错误 (Invalid request data)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "图书未找到或无库存 (Book not found or unavailable)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "数据库错误 (Database error)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/borrow/return": {
            "post": {
                "description": "用户归还图书，更新记录与库存，计算逾期罚款 (User returns a borrowed book, updates record and stock, calculates fine)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "借阅管理 (Borrow Management)"
                ],
                "summary": "归还图书 (User returns a book)",
                "parameters": [
                    {
                        "description": "归还请求参数 (Return Request)",
                        "name": "returnRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BorrowRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "归还成功 (Success Response)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "请求数据错误 (Invalid request data)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "无有效借阅记录 (No active borrow record found)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "数据库错误 (Database error)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/reviews": {
            "get": {
                "description": "获取指定图书的所有用户书评，包括评分、内容和评论者 (Get all reviews for a specific book, including rating, comment, and reviewer)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "书评 (Review)"
                ],
                "summary": "获取图书书评列表 (Retrieve reviews for a book)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "图书 ID (Book ID)",
                        "name": "bookId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "书评列表 (List of reviews)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "400": {
                        "description": "缺少图书 ID (Book ID is required)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "数据库错误 (Database error)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "用户对已借阅的图书添加书评，支持评分和评论内容 (User adds a review to a book they have borrowed; includes rating and comment)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "书评 (Review)"
                ],
                "summary": "添加图书书评 (Create a new review)",
                "parameters": [
                    {
                        "description": "书评请求参数 (Review Request)",
                        "name": "review",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "添加成功 (Review added successfully)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "请求数据无效 (Invalid request data)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "仅可评论借阅过的书籍 (Only borrowed books can be reviewed)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "重复评论 (Already reviewed this book)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "数据库或事务错误 (Database or transaction error)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.BorrowRequest": {
            "type": "object",
            "properties": {
                "bookId": {
                    "description": "图书 ID",
                    "type": "integer",
                    "example": 101
                },
                "userId": {
                    "description": "用户 ID",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "models.ReviewRequest": {
            "type": "object",
            "properties": {
                "bookId": {
                    "description": "图书 ID",
                    "type": "integer",
                    "example": 101
                },
                "comment": {
                    "description": "评论内容",
                    "type": "string",
                    "example": "Great book!"
                },
                "rating": {
                    "description": "评分（1-5）",
                    "type": "integer",
                    "example": 5
                },
                "userId": {
                    "description": "用户 ID",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    }
}